+++ app/admin/product-menu/page.tsx
@@ -1,5 +1,0 @@
-import ProductMenuBuilder from "./ProductMenuBuilder";
-
-export default function Page() {
-  return <ProductMenuBuilder />;
-}